Modified the MakefileWorker so that dual targetting is slightly easier.
Added in a TARGET_PLATFORM variable so that the outputs can easily be
redirected into subdirectories when using different compilers.
Specifically, this will cause CPPUTEST_OBJS_DIR to be set to
objs/$(TARGET_PLATFORM) and CPPUTEST_LIB_DIR to lib/$(TARGET_PLATFORM).
This also modifies the path used for CPPUTEST_LIB so that the correct
library is used to link against.