Suppress Clang 3.6 warnings
The memory leak detector overrides the c++ new keyword with a define and
in Clang 3.6 a 'keyword-macro' warning is issued. Suppress the warning as
this is the intended functionality.

A 'reserved-id-macro' warning is issued for defines starting with __, given
that some of these defines are being created to make Clang compatible with
the other compilers, suppress the warning for those defines.