Optimized MockExpectedCallsList::isEmpty().
This simple modification changes the order of this function to O(1)
instead of O(n).

Taking in account that this function is called at least once each time
that an actual call is created, or withParameter()/onObject() is called
on it, the overall actual calls processing can be reduced from O(n!) to
O(n)!