Moving the test source files to fix error introduced with 02513e803535 (pull request #1059)
There is a problem with dependencies: "AllTests.o" will be different between CppUTestTest and CppUTestExtTest, which is not handled correctly in IAR Embedded Workbench 7.70.1. A full rebuild is needed when trying to run the tests, as the last built AllTests.o is used when linking regardless of which project it was built for.

Interestingly enough, this problem only occurs within Embedded Workbench. When running the scripted tests from within platforms\iar\settings there are no problems. (If the scripts are missing, try to open each test project from within Embedded Workbench, then they should be created)