move related MockParameterTests together
Combined all of the mismatched long long and other type tests into one
block. This reduces the number of checks for CPPUTEST_USE_LONG_LONG, and
hopefully makes the code easier to read.

I chose not to move the expectOne*ParameterAndValue tests, because these
are not related to the mismatched type tests, and I wanted to keep
similar tests together. Moving them would make the test layout a bit
more confusing.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>