Create src/Platforms/Borland/UtestPlatforms.cpp
This is currently a direct copy of src/Platforms/Gcc/UtestPlatforms.cpp.   I have to start with something and this file mostly works.  I will follow up with the required changes latter.

As per the first reply in #1493 this is the first step in getting a working Borland/UtestPlatforms.cpp